Add StopListViewModel
